requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
Pygments==2.18.0
requests==2.32.3
rich==13.9.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
